home page, nodes crud: mynodes/nodes, login

+++ resources/views/nodes/index.blade.php
@@ -1,7 +1,7 @@
 @extends('layout.admin')
 
 @section('title')
-    <a href="#" class="mobile-tittle">Nodes</a>
+    <a href="#" class="mobile-tittle">All Nodes</a>
 @stop
 
 
@@ -34,23 +34,30 @@
                         <td>{{ $node->latitude }}</td>
                         <td>
                             <a href="{{url('nodes',$node->id)}}"><i class="material-icons">visibility</i></a>
-                            <a href="{{route('nodes.edit', $node->id)}}"><i class="material-icons">edit</i></a>
-                            <a class="modal-trigger" href="#modal2"><i class="material-icons">delete</i></a>
+
+                            @if(Auth::check() and Auth::user()->role == 'superadmin')
+                                <a href="{{route('nodes.edit', $node->id)}}"><i class="material-icons">edit</i></a>
+                            @endif
+
+                            @if(Auth::check() and Auth::user()->role == 'superadmin')
+                                <a class="modal-trigger" href="#modal2"><i class="material-icons">delete</i></a>
+
+                                <!-- cancel modal Structure -->
+                                <div id="modal2" class="modal">
+                                    <div class="modal-content center">
+                                        <h6 class="light">This action can not be reversed, would you like to continue? </h6><br>
+                                        <div class="modal-footer">
+                                            {!! Form::open(['method' => 'DELETE', 'route'=>['nodes.destroy', $node->id]]) !!}
+                                            <button type="submit" class="btn btn-flat">Yes</button>
+                                            {!! Form::close() !!}
+                                            <button class="btn btn-flat modal-action modal-close">No</button>
+                                        </div>
+                                    </div>
+                                </div>
+                            @endif
+
                         </td>
                     </tr>
-
-                    <!-- cancel modal Structure -->
-                    <div id="modal2" class="modal">
-                        <div class="modal-content center">
-                            <h6 class="light">This action can not be reversed, would you like to continue? </h6><br>
-                            <div class="modal-footer">
-                                {!! Form::open(['method' => 'DELETE', 'route'=>['nodes.destroy', $node->id]]) !!}
-                                <button type="submit" class="btn btn-flat">Yes</button>
-                                {!! Form::close() !!}
-                                <button class="btn btn-flat modal-action modal-close">No</button>
-                            </div>
-                        </div>
-                    </div>
                 @endforeach
                 </tbody>
             </table>
@@ -58,10 +65,10 @@
     </div>
 
     <!-- FLOATING BUTTON -->
-    <div class="fixed-action-btn" id="add">
+    <!--<div class="fixed-action-btn" id="add">
         <a href="{{url('/nodes/create')}}" class="btn-floating btn-large waves-effect waves-circle waves-light red">
             <i class="large material-icons">add</i>
         </a>
-    </div>
+    </div>-->
     <!-- FLOATING BUTTON -->
 @endsection
